Add init subcommand and allow re-initialization
This commit adds a new `init` subcommand to the CLI for creating a new Cazan project. This includes the creation of a new directory named `.cazan` in the current working directory, along with some default files. Furthermore, a `--re-init` flag has also been introduced which will allow for the re-initialization of an existing Cazan project. This means an existing `.cazan` directory and all of its contents can now be overwritten.